Extra

+++ controller/userController.js
@@ -1,4 +1,4 @@
-const { User } = require('../models');
+const { User, Thought} = require('../models');
 
 module.exports = {
     // Get all users
@@ -40,9 +40,20 @@ module.exports = {
       },
       
   
-    // Delete a user by id
-    deleteUser({ params }, res) {
-      User.findOneAndDelete({ _id: params.id })
+// Delete a user by id and their associated thoughts
+deleteUser({ params }, res) {
+  User.findOne({ _id: params.id })
+    .then((user) => {
+      if (!user) {
+        return res.status(404).json({ message: 'No user with this id!' });
+      }
+
+      // Delete user's thoughts
+      Thought.deleteMany({ username: user.username })
+        .then(() => {
+          // Delete user
+          return User.findOneAndDelete({ _id: params.id });
+        })
         .then((user) => {
           if (!user) {
             return res.status(404).json({ message: 'No user with this id!' });
@@ -50,7 +61,10 @@ module.exports = {
           return res.json(user);
         })
         .catch((err) => res.status(400).json(err));
-    },
+    })
+    .catch((err) => res.status(400).json(err));
+},
+
   
     // Add a friend to a user's friend list
     addFriend({ params, body }, res) {
